+++ pynonymizer/cli.py
@@ -162,6 +162,8 @@ def cli(rawArgs=None):
         logger.warning("Positional STRATEGYFILE is deprecated. Use the -s/--strategy option instead.")
     if args.legacy_output:
         logger.warning("Positional OUTPUT is deprecated. Use the -o/--output option instead.")
+    if args.fake_locale:
+        logger.warning("The locale option -l/--fake-locale is deprecated. Use the locale: key in your strategyfile instead.")
 
     _warn_deprecated_env("DB_TYPE", "PYNONYMIZER_DB_TYPE")
     _warn_deprecated_env("DB_HOST", "PYNONYMIZER_DB_HOST")
@@ -172,7 +174,10 @@ def cli(rawArgs=None):
     
     if args.verbose:
         console_handler.setLevel(logging.DEBUG)
-        
+
+    # Add local project dir to path in case of custom provider imports     
+    if "." not in sys.path:
+        sys.path.append(".")
     try:
         pynonymize(
             input_path=input,